+++ controllers/workspace.controller.js
@@ -28,7 +28,7 @@ class WorkspaceController {
           message: err.message,
         });
       }
-      return ErrorUtils.handleInternalServerError(res);
+      return ErrorUtils.handleUnexpectedError(res);
     }
   };
 
@@ -44,7 +44,7 @@ class WorkspaceController {
           message: err.message,
         });
       }
-      return ErrorUtils.handleInternalServerError(res);
+      return ErrorUtils.handleUnexpectedError(res);
     }
   };
 
@@ -52,12 +52,22 @@ class WorkspaceController {
   getOneWorkspace = async (req, res, next) => {
     try {
       const workspaceId = req.params.workspaceId;
-
+      
       // 입력 데이터에 대한 유효성 검사
       if (!workspaceId) {
         throw new ErrorUtils(
           StatusCodes.BAD_REQUEST,
           "workspaceId는 필수 입력값입니다."
+        );
+      }
+
+      // params의 값이 objectId의 형식에 맞는지 검사
+      // 24개의 문자로 구성된 16진수 문자열 인지 확인
+      const isValidHex = /^[0-9a-fA-F]{24}$/.test(workspaceId);
+      if (!isValidHex) {
+        throw new ErrorUtils(
+          StatusCodes.BAD_REQUEST,
+          "workspaceId가 유효하지 않습니다."
         );
       }
 
@@ -72,7 +82,7 @@ class WorkspaceController {
           message: err.message,
         });
       }
-      return ErrorUtils.handleInternalServerError(res);
+      return ErrorUtils.handleUnexpectedError(res);
     }
   };
 
@@ -90,6 +100,16 @@ class WorkspaceController {
         );
       }
 
+      // params의 값이 objectId의 형식에 맞는지 검사
+      // 24개의 문자로 구성된 16진수 문자열 인지 확인
+      const isValidHex = /^[0-9a-fA-F]{24}$/.test(workspaceId);
+      if (!isValidHex) {
+        throw new ErrorUtils(
+          StatusCodes.BAD_REQUEST,
+          "workspaceId가 유효하지 않습니다."
+        );
+      }
+
       await this.workspaceService.putUserToWorkspace(workspaceId, newMember);
       return res.status(StatusCodes.CREATED).end();
     } catch (err) {
@@ -99,7 +119,7 @@ class WorkspaceController {
           message: err.message,
         });
       }
-      return ErrorUtils.handleInternalServerError(res);
+      return ErrorUtils.handleUnexpectedError(res);
     }
   };
 
@@ -116,6 +136,16 @@ class WorkspaceController {
         );
       }
 
+      // params의 값이 objectId의 형식에 맞는지 검사
+      // 24개의 문자로 구성된 16진수 문자열 인지 확인
+      const isValidHex = /^[0-9a-fA-F]{24}$/.test(workspaceId);
+      if (!isValidHex) {
+        throw new ErrorUtils(
+          StatusCodes.BAD_REQUEST,
+          "workspaceId가 유효하지 않습니다."
+        );
+      }
+
       await this.workspaceService.deleteWorkspace(workspaceId);
       return res.status(StatusCodes.NO_CONTENT).end();
     } catch (err) {
@@ -125,7 +155,7 @@ class WorkspaceController {
           message: err.message,
         });
       }
-      return ErrorUtils.handleInternalServerError(res);
+      return ErrorUtils.handleUnexpectedError(res);
     }
   };
 }
