+++ services/channel.service.js
@@ -1,3 +1,4 @@
+const { StatusCodes } = require("http-status-codes");
 const ErrorUtils = require("../utils/error.utils");
 const ChannelRepository = require("../repositories/channel.repository");
 const WorkspaceRepository = require("../repositories/workspace.repository");
@@ -8,20 +9,52 @@ class ChannelService {
 
   // Channel 생성
   postChannel = async (channelName, userName, workspaceId, res, next) => {
+    // 워크스페이스 조회
+    let workspace = await this.workspaceRepository.getOneWorkspaceById( workspaceId );
+    
+    // 해당 워크스페이스가 존재하지 않음
+    if (!workspace) {
+      throw new ErrorUtils(
+        StatusCodes.NOT_FOUND,
+        "워크스페이스가 존재하지 않습니다."
+      );
+    };
+
+    // 워크스페이스 채널에 이미 해당 채널이 있음
+    if (workspace.workspaceChannel.includes(channelName)) {
+      throw new ErrorUtils(
+        StatusCodes.BAD_REQUEST,
+        `${channelName}은/는 이미 해당 워크스페이스의 채널입니다.`
+      );
+    }
+    
     // 워크스페이스에 채널 정보 반영
-    const workspace = await this.workspaceRepository.putChannelToWorkspace(
+    workspace = await this.workspaceRepository.putChannelToWorkspace(
       workspaceId,
       channelName
     );
+
+    // 워크스페이스에 채널 정보 반영 실패
     if (!workspace) {
       throw new ErrorUtils(
         StatusCodes.BAD_REQUEST,
-        "워크스페이스에 채널 정보 반영 실패"
+        "워크스페이스에 채널 정보 반영에 실패했습니다."
+      );
+    }
+
+    // 채널 조회
+    let channel = await this.channelRepository.getOneChannelbyName( channelName, workspaceId );
+
+    // 이미 해당 워크스페이스와 채널 정보를 가진 값이 존재
+    if (channel) {
+      throw new ErrorUtils(
+        StatusCodes.BAD_REQUEST,
+        "이미 해당 워크스페이스와 채널 정보를 가진 값이 존재합니다."
       );
     }
 
     // 채널 생성
-    const channel = await this.channelRepository.postChannel(
+    channel = await this.channelRepository.postChannel(
       channelName,
       workspaceId,
       userName
@@ -29,7 +62,10 @@ class ChannelService {
 
     // 채널이 정상적으로 생성되지 않음
     if (!channel) {
-      throw new ErrorUtils(StatusCodes.BAD_REQUEST, "채널 생성 실패");
+      throw new ErrorUtils(
+        StatusCodes.BAD_REQUEST,
+        "채널 생성에 실패하였습니다."
+      );
     }
 
     return channel;
@@ -40,33 +76,45 @@ class ChannelService {
     const workspace = await this.workspaceRepository.putChannelToWorkspace(
       workspaceId
     );
+
+    // 워크스페이스가 존재하지 않음
     if (!workspace) {
       throw new ErrorUtils(
-        StatusCodes.BAD_REQUEST,
-        "워크스페이스에 채널 정보 반영 실패"
+        StatusCodes.NOT_FOUND,
+        "워크스페이스가 존재하지 않습니다."
       );
     }
 
-    const channelList = await this.channelRepository.getAllChannel(workspaceId);
+    const channel = await this.channelRepository.getAllChannel(workspaceId);
 
-    const filteredChannelList = channelList.map((channel) => ({
-      channelName: channel.channelName,
-      channelId: channel.channelId,
-      userCount: channel.channelMember.length,
+    const filteredChannel = channel.map((item) => ({
+      channelName: item.channelName,
+      channelId: item.channelId,
+      userCount: item.channelMember.length,
     }));
 
-    return filteredChannelList;
+    return filteredChannel;
   };
 
   // 특정 Channel 정보 가저오기
   getOneChannel = async (workspaceId, channelId, res, next) => {
-    const channel = await this.channelRepository.getOneChannel(channelId);
+    const workspace = await this.workspaceRepository.putChannelToWorkspace(workspaceId);
 
-    // 해당 채널이 없는 경우 에러
+    // 워크스페이스가 존재하지 않음
+    if (!workspace) {
+      throw new ErrorUtils(
+        StatusCodes.NOT_FOUND,
+        "워크스페이스가 존재하지 않습니다."
+      );
+    }
+    
+    const channel = await this.channelRepository.getOneChannelbyId(channelId);
+
+    // 채널이 없는 경우 에러
     if (!channel) {
       throw new ErrorUtils(
         StatusCodes.BAD_REQUEST,
-        "해당 채널이 존재하지 않습니다."
+        "채널이 존재하지 않습니다."
       );
     }
 
@@ -81,17 +129,65 @@ class ChannelService {
 
   // Channel 수정 - 인원 추가
   putUserToChannel = async (workspaceId, channelId, newMember, res, next) => {
-    const channel = await this.channelRepository.putUserToChannel(
+    const workspace = await this.workspaceRepository.putChannelToWorkspace(workspaceId);
+
+    // 워크스페이스가 존재하지 않음
+    if (!workspace) {
+      throw new ErrorUtils(
+        StatusCodes.NOT_FOUND,
+        "워크스페이스가 존재하지 않습니다."
+      );
+    }
+    
+    let channel = await this.channelRepository.getOneChannelbyId(channelId);
+
+    // 채널이 없는 경우 에러
+    if (!channel) {
+      throw new ErrorUtils(
+        StatusCodes.BAD_REQUEST,
+        "채널이 존재하지 않습니다."
+      );
+    }
+
+    // 워크스페이스 멤버에 이미 해당 인원이 있음
+    if (channel.channelMember.includes(newMember)) {
+      throw new ErrorUtils(
+        StatusCodes.BAD_REQUEST,
+        `${newMember}은/는 이미 해당 채널의 멤버입니다.`
+      );
+    }
+
+    channel = await this.channelRepository.putUserToChannel(
       channelId,
       newMember
     );
-    return channel;
+    // return channel;
   };
 
   // Channel 삭제
   deleteChannel = async (workspaceId, channelId, res, next) => {
-    const channel = await this.channelRepository.deleteChannel(channelId);
-    return channel;
+    const workspace = await this.workspaceRepository.putChannelToWorkspace(workspaceId);
+
+    // 워크스페이스가 존재하지 않음
+    if (!workspace) {
+      throw new ErrorUtils(
+        StatusCodes.NOT_FOUND,
+        "워크스페이스가 존재하지 않습니다."
+      );
+    }
+    
+    let channel = await this.channelRepository.getOneChannelbyId(channelId);
+
+    // 채널이 없는 경우 에러
+    if (!channel) {
+      throw new ErrorUtils(
+        StatusCodes.BAD_REQUEST,
+        "채널이 존재하지 않습니다."
+      );
+    }
+    
+    await this.channelRepository.deleteChannel(channelId);
+    // return channel;
   };
 }
